Rename _native to native and re-export from top-level package
Plugin authors doing a deep integration need the rust extension's
ProjectContext, GraphOp constructors, and the query() builder. The
underscore-prefixed dead_cst._native module signaled "private" and
forced them through a private path. Rename to dead_cst.native, ship
the wheel as python/dead_cst/native.{abi3.so,pyd} + native.pyi, and
re-export it from the top-level package so the supported idiom is
``from dead_cst import native``. No compat shim — dead-cst is alpha.
