+++ python/dead_cst/__init__.py
@@ -1,7 +1,7 @@
 """Dead code analysis for Python.
 
 ``dead-cst`` builds a symbol-level reachability graph of a Python
-codebase via a rust-backed extension (:mod:`dead_cst._native`, which
+codebase via a rust-backed extension (:mod:`dead_cst.native`, which
 uses ty's ``SemanticIndex``) and reports (or removes) anything not
 reachable from a configurable set of entrypoints.
 
@@ -21,12 +21,17 @@ The deeper public surface lives in focused sub-packages:
 * :mod:`dead_cst.resolvers` -- the :class:`PathResolver` protocol,
   :class:`ManualResolver`, :class:`UvResolver`.
 * :mod:`dead_cst.contrib` -- third-party-aware extensions.
+* :mod:`dead_cst.native` -- the rust-backed extension. Plugin authors
+  doing a deep integration import this directly
+  (``from dead_cst import native``) to construct ``GraphOp``\\s and
+  query the in-progress :class:`native.ProjectContext`.
 
 ``dead-cst`` is alpha; APIs, CLI flags, and output formats may change.
 """
 
 from importlib.metadata import PackageNotFoundError, version as _pkg_version
 
+from . import native
 from .analyze import Analysis, PackageView
 from .graph import EdgeFlags, Import, NodeFlags, SymbolNode
 
@@ -43,4 +48,5 @@ __all__ = [
     "PackageView",
     "SymbolNode",
     "__version__",
+    "native",
 ]
