+++ python/dead_cst/native.pyi
