+++ python/dead_cst/plugins/__init__.py
@@ -1,7 +1,7 @@
 """Pluggable edge contributors.
 
 A plugin subclasses :class:`Plugin` and implements ``run(ctx)``
-against a :class:`dead_cst._native.ProjectContext`, yielding
+against a :class:`dead_cst.native.ProjectContext`, yielding
 :class:`AddNode` / :class:`AddEdge` / :class:`AddEntrypoint` ops the
 rust backend applies to the in-progress project graph.
