+++ python/dead_cst/plugins/_base.py
@@ -4,7 +4,7 @@ Every concrete plugin subclasses :class:`Plugin`. The base exists so
 the rust backend can do an :func:`isinstance` check (typos like
 ``Pluign()`` raise :class:`TypeError` instead of being silently
 skipped) and so plugin authors don't have to repeat the
-``from dead_cst import _native as native`` lazy import at the top of
+``from dead_cst import native`` lazy import at the top of
 every ``run()`` body.
 
 The base is intentionally tiny — it does not enforce the ``run``
@@ -14,7 +14,7 @@ signature today. Concrete plugins are expected to expose
 
 from __future__ import annotations
 
-from dead_cst import _native as native
+from dead_cst import native
 
 __all__ = ["Plugin", "native"]
 
@@ -27,9 +27,8 @@ class Plugin:
     The base class itself does nothing — it just provides a runtime
     type the backend can check.
 
-    Access the native extension as ``self.native`` (or via the
-    re-export from :mod:`dead_cst.plugins._base`) instead of repeating
-    ``from dead_cst import _native as native`` inside every ``run()``.
+    Access the native extension as ``self.native`` instead of
+    repeating ``from dead_cst import native`` inside every ``run()``.
     """
 
     native = native
