+++ tests/prototype/test_native_graph.py
@@ -18,7 +18,7 @@ from pathlib import Path
 
 import pytest
 
-native = pytest.importorskip("dead_cst._native")
+native = pytest.importorskip("dead_cst.native")
 
 from dead_cst.graph import EdgeFlags, NodeFlags  # noqa: E402
